Mms: Avoid showing a window when marking as read
Change the "mark as read" action implementation, so that it
uses a BroadcastReceiver instead of an Activity.
The downside of using an Activity is that the current top
activity will be paused and resumed. This pausing, however
brief, may disturb what the user is doing, like watching a
kitty video on youtube.

Signed-off-by: ThiagoVinicius <thiagovfar@gmail.com>